+++ packages/eslint-plugin/tests/rules/directive-selector/cases.ts
@@ -181,6 +181,101 @@ export const valid: readonly (string | ValidTestCase<Options>)[] = [
       },
     ],
   },
+  // Single config array - element only
+  {
+    code: `
+      @Directive({
+        selector: 'app-foo-bar'
+      })
+      class Test {}
+      `,
+    options: [[{ type: 'element', prefix: 'app', style: 'kebab-case' }]],
+  },
+  // Single config array - attribute only
+  {
+    code: `
+      @Directive({
+        selector: '[appFooBar]'
+      })
+      class Test {}
+      `,
+    options: [[{ type: 'attribute', prefix: 'app', style: 'camelCase' }]],
+  },
+  // Multiple configs - element with kebab-case
+  {
+    code: `
+      @Directive({
+        selector: 'app-foo-bar'
+      })
+      class Test {}
+      `,
+    options: [
+      [
+        { type: 'element', prefix: 'app', style: 'kebab-case' },
+        { type: 'attribute', prefix: 'app', style: 'camelCase' },
+      ],
+    ],
+  },
+  // Multiple configs - attribute with camelCase
+  {
+    code: `
+      @Directive({
+        selector: '[appFooBar]'
+      })
+      class Test {}
+      `,
+    options: [
+      [
+        { type: 'element', prefix: 'app', style: 'kebab-case' },
+        { type: 'attribute', prefix: 'app', style: 'camelCase' },
+      ],
+    ],
+  },
+  // Multiple configs - element with array of prefixes
+  {
+    code: `
+      @Directive({
+        selector: 'lib-foo-bar'
+      })
+      class Test {}
+      `,
+    options: [
+      [
+        { type: 'element', prefix: ['app', 'lib'], style: 'kebab-case' },
+        { type: 'attribute', prefix: 'app', style: 'camelCase' },
+      ],
+    ],
+  },
+  // Multiple configs - attribute with array of prefixes
+  {
+    code: `
+      @Directive({
+        selector: '[libFooBar]'
+      })
+      class Test {}
+      `,
+    options: [
+      [
+        { type: 'element', prefix: 'app', style: 'kebab-case' },
+        { type: 'attribute', prefix: ['app', 'lib'], style: 'camelCase' },
+      ],
+    ],
+  },
+  // Multiple configs - config order shouldn't matter
+  {
+    code: `
+      @Directive({
+        selector: '[appFooBar]'
+      })
+      class Test {}
+      `,
+    options: [
+      [
+        { type: 'attribute', prefix: 'app', style: 'camelCase' },
+        { type: 'element', prefix: 'app', style: 'kebab-case' },
+      ],
+    ],
+  },
 ];
 
 export const invalid: readonly InvalidTestCase<MessageIds, Options>[] = [
@@ -291,5 +386,109 @@ export const invalid: readonly InvalidTestCase<MessageIds, Options>[] = [
     messageId: messageIdTypeFailure,
     options: [{ type: 'element', prefix: ['app', 'ng'], style: 'camelCase' }],
     data: { type: 'element' },
+  }),
+  // Single config array - element with wrong style
+  convertAnnotatedSourceToFailureCase({
+    description: `should fail if an element selector doesn't match when using single config array`,
+    annotatedSource: `
+        @Directive({
+          selector: 'appFooBar'
+                    ~~~~~~~~~~~
+        })
+        class Test {}
+      `,
+    messageId: messageIdStyleFailure,
+    options: [[{ type: 'element', prefix: 'app', style: 'kebab-case' }]],
+    data: { style: 'kebab-case' },
+  }),
+  // Single config array - attribute with wrong style
+  convertAnnotatedSourceToFailureCase({
+    description: `should fail if an attribute selector doesn't match when using single config array`,
+    annotatedSource: `
+        @Directive({
+          selector: '[app-foo-bar]'
+                    ~~~~~~~~~~~~~~~
+        })
+        class Test {}
+      `,
+    messageId: messageIdStyleFailure,
+    options: [[{ type: 'attribute', prefix: 'app', style: 'camelCase' }]],
+    data: { style: 'camelCase' },
+  }),
+  // Multiple configs - element with wrong style
+  convertAnnotatedSourceToFailureCase({
+    description: `should fail if an element selector doesn't match kebab-case when using multiple configs`,
+    annotatedSource: `
+        @Directive({
+          selector: 'appFooBar'
+                    ~~~~~~~~~~~
+        })
+        class Test {}
+      `,
+    messageId: messageIdStyleFailure,
+    options: [
+      [
+        { type: 'element', prefix: 'app', style: 'kebab-case' },
+        { type: 'attribute', prefix: 'app', style: 'camelCase' },
+      ],
+    ],
+    data: { style: 'kebab-case' },
+  }),
+  // Multiple configs - attribute with wrong style
+  convertAnnotatedSourceToFailureCase({
+    description: `should fail if an attribute selector doesn't match camelCase when using multiple configs`,
+    annotatedSource: `
+        @Directive({
+          selector: '[app-foo-bar]'
+                    ~~~~~~~~~~~~~~~
+        })
+        class Test {}
+      `,
+    messageId: messageIdStyleFailure,
+    options: [
+      [
+        { type: 'element', prefix: 'app', style: 'kebab-case' },
+        { type: 'attribute', prefix: 'app', style: 'camelCase' },
+      ],
+    ],
+    data: { style: 'camelCase' },
+  }),
+  // Multiple configs - element with wrong prefix
+  convertAnnotatedSourceToFailureCase({
+    description: `should fail if an element selector has wrong prefix when using multiple configs`,
+    annotatedSource: `
+        @Directive({
+          selector: 'lib-foo-bar'
+                    ~~~~~~~~~~~~~
+        })
+        class Test {}
+      `,
+    messageId: messageIdPrefixFailure,
+    options: [
+      [
+        { type: 'element', prefix: 'app', style: 'kebab-case' },
+        { type: 'attribute', prefix: 'app', style: 'camelCase' },
+      ],
+    ],
+    data: { prefix: '"app"' },
+  }),
+  // Multiple configs - attribute with wrong prefix
+  convertAnnotatedSourceToFailureCase({
+    description: `should fail if an attribute selector has wrong prefix when using multiple configs`,
+    annotatedSource: `
+        @Directive({
+          selector: '[libFooBar]'
+                    ~~~~~~~~~~~~~
+        })
+        class Test {}
+      `,
+    messageId: messageIdPrefixFailure,
+    options: [
+      [
+        { type: 'element', prefix: 'app', style: 'kebab-case' },
+        { type: 'attribute', prefix: 'app', style: 'camelCase' },
+      ],
+    ],
+    data: { prefix: '"app"' },
   }),
 ];